Add sensor_config.h to avoid duplicate structure definitions
- Created sensor_config.h with SensorConfig structure
- Updated ClassFlowSensors.h and sensor_manager.h to include it
- Removed duplicate SensorConfig definition from sensor_manager.cpp
