+++ code/components/jomjol_flowcontroll/ClassFlowSensors.h
@@ -4,25 +4,12 @@
 #define CLASSFLOWSENSORS_H
 
 #include "ClassFlow.h"
+#include "../jomjol_sensors/sensor_config.h"
 #include <memory>
 #include <map>
 
 class SensorManager;
 class ClassFlowControll;
-
-// Configuration structure for sensors
-struct SensorConfig {
-    bool enable = false;
-    int interval = -1;  // -1 = follow flow (default)
-    bool mqttEnable = true;
-    std::string mqttTopic;
-    bool influxEnable = false;
-    std::string influxMeasurement;
-    
-    // SHT3x specific
-    uint8_t sht3xAddress = 0x44;
-    uint32_t i2cFreq = 100000;
-};
 
 class ClassFlowSensors : public ClassFlow
 {
